+++ src/app/src/components/ClaimFacilityIntroStep.jsx
@@ -10,22 +10,31 @@ import { claimAFacilityFormStyles } from '../util/styles';
 
 import COLOURS from '../util/COLOURS';
 
-const confirmationMessageStyles = Object.freeze({
+const introMessageStyles = Object.freeze({
     width: '95%',
+    padding: '10px',
 });
 
 const checkIconStyles = Object.freeze({
     color: COLOURS.NAVY_BLUE,
 });
 
-export default function ClaimFacilityConfirmationStep() {
+export default function ClaimFacilityIntroStep() {
     return (
         <div style={claimAFacilityFormStyles.inputGroupStyles}>
-            <Typography variant="headline" style={confirmationMessageStyles}>
-                Your request to claim the facility was submitted successfully!
-                To manage the profile information for your facility, we&#39;ll
-                need to verify your connection with the facility. Once verified
-                you will be able to:
+            <Typography variant="headline" style={introMessageStyles}>
+                Owners or senior management at facilities listed on the OAR are
+                able to &quot;claim&quot; their facility&#39;s profile on the site and add
+                business information to it.
+            </Typography>
+            <Typography variant="headline" style={introMessageStyles}>
+                If you are a facility owner or senior management, please work
+                your way through the following three preliminary steps to
+                verify your identity. This process should take no more than 5
+                minutes to complete.
+            </Typography>
+            <Typography variant="headline" style={introMessageStyles}>
+                Once your claim has been reviewed and approved, you will be able to:
             </Typography>
             <List>
                 <ListItem>
@@ -64,12 +73,14 @@ export default function ClaimFacilityConfirmationStep() {
                     </ListItemIcon>
                     <ListItemText>
                         <Typography variant="headline">
-                            Add information about your head office and parent
-                            company
+                            Add information about your head office and parent company
                         </Typography>
                     </ListItemText>
                 </ListItem>
             </List>
+            <Typography variant="title" style={introMessageStyles}>
+                This information will be shown publicly on the facility details page.
+            </Typography>
         </div>
     );
 }